Refactor character module structure and implement new classes
- Deleted the old character classes module and created a new models package.
- Introduced a new attributes class to manage character attributes.
- Created an interface for character classes with abstract methods.
- Implemented specific character classes: Barbarian, Fighter, Scholar, and Wizard.
- Developed race classes: Human, Elf, Dwarf, Halfling, Gnome, and Half-Elf with their respective attributes and features.
- Added rolling styles for character generation: Classic, Adventurous, and Heroic.
- Updated the character service to utilize the new models and classes.
