--- src/dragon/character/classes/__init__.py
@@ -1,6 +1,0 @@
-"""Módulo de classes de personagem.
-
-Use imports explícitos:
-	from dragon.character.classes.figher import Fighter
-	from dragon.character.data.combat_style import CombatStyle
-"""
