+++ src/dragon/models/__init__.py
@@ -1,5 +1,5 @@
 """Pacote character: não reexporta mais raças ou classes para evitar acoplamento.
 
 Importe explicitamente de submódulos, por exemplo:
-	from dragon.character.races.human import Human
+        from dragon.models.races.human import Human
 """
