+++ src/dragon/models/classes/__init__.py
@@ -1,0 +1,6 @@
+"""Módulo de classes de personagem.
+
+Use imports explícitos:
+        from dragon.models.classes.figher import Fighter
+        from dragon.models.data.combat_style import CombatStyle
+"""
