+++ src/dragon/models/classes/barbarian.py
@@ -1,6 +1,7 @@
 from typing import List
 from .IClasses import ICharacterClass
-from dragon.character.data.live_dice import LiveDice
+from dragon.models.data.live_dice import LiveDice
+
 
 class Barbarian(ICharacterClass):
     """Bárbaro é um combatente selvagem."""
