+++ src/dragon/models/classes/scholar.py
@@ -1,6 +1,6 @@
 from typing import List
 from .IClasses import ICharacterClass
-from dragon.character.data.live_dice import LiveDice
+from dragon.models.data.live_dice import LiveDice
 
 
 class Scholar(ICharacterClass):
